+++ MsgServer/MsgServerSequence/SRTResponseCollection.h
@@ -21,35 +21,33 @@
 #define REQUEST_TYPE_WRITE (2)
 
 #define DEF_PROTO 1
-#include "MsgServer/proto/storage_msg.pb.h"
-#include "MsgServer/proto/storage_msg_type.pb.h"
+#include "ProtoCommon.h"
 
 class SRTRedisManager;
 
 class SRTResponseCollection {
 public:
-    SRTResponseCollection(SRTRedisManager* manager, int reqType, int clientNum, const pms::StorageMsg& request, long long seqn);
+    SRTResponseCollection(SRTRedisManager* manager, int reqType, int clientNum, const pms::StorageMsg& request, int64 seqn);
     ~SRTResponseCollection();
 
-    void AddResponse(const pms::StorageMsg& request, long long seqn);
-    //const std::string& GetUserid() { return m_UserId; }
+    void AddResponse(const pms::StorageMsg& request, int64 seqn);
 
     // for all the sequence request
     // read request update maxseqn()
     // write request udpate sequence()
     class MsgSeqn {
         public:
-        MsgSeqn(int clientNum, long long seqn, SRTRedisManager* manager, const pms::StorageMsg& request);
+        MsgSeqn(int clientNum, int64 seqn, SRTRedisManager* manager, const pms::StorageMsg& request);
         ~MsgSeqn();
         int                 cnumber;
         int                 counter;
         SRTRedisManager*       rmanager;
-        std::set<long long> seqns;
+        std::set<int64> seqns;
         pms::StorageMsg     storage;
         //pms::SeqnMsg
         sigslot::signal1<const std::string&> WriteResponse;
         sigslot::signal1<const std::string&> ReadResponse;
-        bool AddAndCheckWrite(long long seq)
+        bool AddAndCheckWrite(int64 seq)
         {
             seqns.insert(seq);
             //printf("AddAndCheckWrite counter is:%d, cnumber is:%d\n", counter, cnumber);
@@ -67,7 +65,7 @@ public:
                 return false;
             }
         }
-        bool AddAndCheckRead(long long seq)
+        bool AddAndCheckRead(int64 seq)
         {
             seqns.insert(seq);
             printf("AddAndCheckRead counter is:%d, cnumber is:%d\n", counter, cnumber);
